# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meter API",
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/equipements/user/{id}/compteur/{id}": {
            "get": {
                "tags": [
                    "Equipements"
                ],
                "summary": "Obtenir les equipements d'un utilisateur",
                "operationId": "26c316fcac7a89d1c030c2f38b09b0b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des equipements lié a un utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "schemas/Equipement"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
        "/api/equipements/show/{id}": {
            "get": {
                "tags": [
                    "Equipements"
                ],
                "summary": "Afficher les détails d'un équipement",
                "operationId": "836564cc29fa54c3b587b090877d381d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'équipement à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'équipement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "schemas/Equipement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Équipement non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Cet équipement ne figure pas dans notre base de données"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipements/new": {
            "post": {
                "tags": [
                    "Equipements"
                ],
                "summary": "Enregistrer un nouvel équipement",
                "operationId": "e8a0acc3f994d8f186f261d0041aa6b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom_appareil",
                                    "puissance",
                                    "compteur_id"
                                ],
                                "properties": {
                                    "nom_appareil": {
                                        "type": "string",
                                        "example": "Réfrigérateur"
                                    },
                                    "puissance": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1200.5
                                    },
                                    "compteur_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Équipement enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Succès"
                                        },
                                        "data": {
                                            "$ref": "schemas/Equipement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de l'enregistrement de votre équipement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipements/{id}/edit": {
            "get": {
                "tags": [
                    "Equipements"
                ],
                "summary": "Afficher le formulaire d'édition d'un équipement",
                "operationId": "84bde844fa87e722a15ffede62549974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'équipement à éditer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulaire d'édition de l'équipement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "schemas/Equipement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Équipement non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Équipement non trouvé dans notre base de données"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipements/update{id}": {
            "put": {
                "tags": [
                    "Equipements"
                ],
                "summary": "Modifier un équipement",
                "operationId": "f0e65c34511b282a702d0981c541616b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'équipement à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_appareil": {
                                        "type": "string",
                                        "example": "Nouveau nom de l'appareil"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Équipement modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Votre équipement a été modifié avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Équipement non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "L'équipement est introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/equipements/delete/{id}": {
            "delete": {
                "tags": [
                    "Equipements"
                ],
                "summary": "Supprimer un équipement",
                "operationId": "d0744ea142ec1dcecf0fb70a2799ba33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'équipement à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Équipement supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Équipement supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Équipement non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Votre équipement est introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}